Change prescaler from 64 to 8 and fix some types

+++ E7012E-car-project/MCU_Controller/MCU_Controller.h
@@ -1,5 +1,5 @@
-void throttleControl(int speed);
-void steeringControl(int angle);
+void throttleControl(float speed);
+void steeringControl(float angle);
 
 //peripheral defines
 #define SET_STEERING_PWM_REG OCR1B
